+++ front-end/src/containers/BeginComponent/index.jsx
@@ -1,8 +1,13 @@
-import React from "react";
+import React, { useState } from "react";
 import { Typography, Box, TextField, Button } from "@mui/material";
 
-function EmailField({email, setEmail, handleSubmit}) {
-  
+function EmailField() {
+  const [email, setEmail] = useState("");
+  const handleSubmit = () => {
+    if (email.trim() !== "") {
+      console.log("Submitting email: ", email);
+    }
+  };
   const handleKeyDown = (event) => {
     if (event.key === "Enter") {
       handleSubmit();
@@ -72,34 +77,9 @@ function EmailField({email, setEmail, handleSubmit}) {
   );
 }
 
-
 const BeginComponent = (props) => {
   return (
     <div className="BeginComponent">
-      <Button
-        sx={{
-          display: "flex",
-          flexDirection: "row",
-          justifyContent: "center",
-          alignItems: "center",
-          padding: "0px",
-          gap: "8px",
-
-          position: "absolute",
-          width: "83px",
-          height: "40px",
-          left: "360px",
-          top: "29px",
-
-          background: "#FFFFFF",
-          border: "2px solid #000000",
-          variant: "text",
-          color: "#000000",
-        }}
-        >
-        Back
-        </Button>
-        
       <Typography
         sx={{
           position: "absolute",
@@ -141,7 +121,7 @@ const BeginComponent = (props) => {
       >
         Let&apos;s start with your email
       </Typography>
-      <EmailField email={props.email} setEmail={props.setEmail} handleSubmit={props.handleSubmit}/>
+      <EmailField />
     </div>
   );
 };
